commands: consolidate dynamic-decode helpers into shared module (closes #93)
Moves 15 scale-value decoder helpers that were duplicated across
`src/commands/stake.rs` and `src/commands/subnet.rs` into a new
`src/commands/dynamic_decode.rs` module. Every chain-touching command
that uses `subxt::dynamic::*` can now `use crate::commands::
dynamic_decode::*` instead of copy-pasting the same helpers.

## Scope

Pure refactor. Zero behavior change. Every helper's body is moved
unchanged from its original site, and every original site is
converted into a `use` import.

## What moved

14 helpers originally duplicated across stake.rs and subnet.rs:

- **`compact_u16`, `compact_u64`, `compact_u128`** — extract a
  `Compact<T>`-wrapped integer from a named field, coerced to the
  specified width.
- **`compact_value_to_u128`** — the recursive composite-aware integer
  walker that handles both `Primitive(U128(n))` and
  `Composite(Unnamed([U128(n)]))` (Compact-wrapper) shapes.
- **`value_to_u64`** — thin wrapper on `compact_value_to_u128` with a
  `u64::try_from` coerce. Was a standalone function in stake.rs;
  now delegates to `compact_value_to_u128`.
- **`extract_balance_u128`** — identical in behavior to
  `compact_value_to_u128` (now delegates to it). Kept as a distinct
  name because the original call sites read more naturally as
  "balance" rather than "compact integer"; could be deprecated in a
  future cleanup.
- **`extract_account_id_field`** — pull a 32-byte AccountId out of a
  named field, trying both the wrapped-tuple-struct shape and the
  raw byte array shape.
- **`value_to_32_bytes`** — walk a sequence-of-u8 shape into a
  `[u8; 32]` array.
- **`decode_compact_u8_vec`** — lossy UTF-8 decode of
  `Vec<Compact<u8>>` (Metagraph `name` and `symbol` fields).
- **`pad_or_check<T: Default + Clone>`** — normalize a per-row
  parallel `Vec<T>` against an expected length. Three cases:
  `got == expected` passes through, `got.is_empty()` backfills with
  defaults (for upstream-deprecated fields like `rank`/`trust`/
  `pruning_score` on Metagraph), anything else is a hard error.
- **`walk_account_vec`** — walk a `Vec<AccountId32>` field into
  `Vec<[u8; 32]>`.
- **`walk_compact_u16_vec`, `walk_compact_u64_vec`,
  `walk_compact_u128_vec`** — width-specialized wrappers on
  `walk_compact_numeric_vec`.
- **`walk_compact_numeric_vec<T>`** — generic walker with a
  caller-supplied `coerce: Fn(u128) -> Option<T>` closure.
- **`walk_bool_vec`** — walk a `Vec<bool>` field.

Plus one new convenience helper, `extract_account_id_ss58`, that
pairs `extract_account_id_field` with `AccountId32::from(bytes)
.to_ss58check()`. Currently `#[allow(dead_code)]` because no existing
caller uses it, but it's the natural pair for the raw-bytes extractor
and was going to get written inline in the next chain-touching
command anyway.

## Touched files

- `src/commands/dynamic_decode.rs` — NEW, +317 lines. Every helper
  is `pub(crate)` so it's visible to sibling command modules but
  does not leak out of the `commands` module.
- `src/commands/mod.rs` — added `pub mod dynamic_decode;`.
- `src/commands/stake.rs` — deleted 3 local fn bodies
  (`extract_account_id_field`, `value_to_32_bytes`, `value_to_u64`),
  replaced with `use crate::commands::dynamic_decode::{
  extract_account_id_field, value_to_u64};`. `value_to_32_bytes`
  was not directly called from stake.rs (only via the now-moved
  `extract_account_id_field`), so it's not re-imported.
- `src/commands/subnet.rs` — deleted 14 local fn bodies, replaced
  with a 4-line `use` block importing 12 helpers that subnet's
  command functions still call directly. `format_rao_as_tao` stays
  in subnet.rs because it's a formatter (not a decoder) used only
  there.

## Local gates

- `cargo check`: clean (caught one unused `compact_value_to_u128`
  import in the initial pass, removed)
- `cargo clippy --all-targets --all-features -- -D warnings`: clean
- `cargo test` targeting the decoder paths:
  `parse_stake_info_list_{extracts_rows,rejects_bad_netuid}`,
  `format_rao_{zero,one_tao,one_rao,mixed,large_with_fraction,
  u128_max_does_not_panic}`, `mnemonic_recovery_roundtrip`,
  `seed_recovery_roundtrip` → 10 passed, 0 failed
- `parse_stake_info_list_extracts_rows` in particular exercises the
  moved `extract_account_id_field` and `value_to_u64` helpers
  end-to-end through the stake decoder — green proves the move is
  transparent.

## No dep changes

Zero new deps. 4 files touched, +335/-286 net (removing duplicated
bodies from stake.rs + subnet.rs, adding the consolidated copies).

Closes #93.
