+++ src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod chain;
+pub mod dynamic_decode;
 pub mod password_file;
 pub mod paths;
 pub mod skill;
