+++ src/commands/subnet.rs
@@ -14,6 +14,11 @@ use sp_core::crypto::{AccountId32, Ss58Codec};
 use subxt::dynamic::{At, Value};
 use subxt::ext::scale_value::Value as SValue;
 
+use crate::commands::dynamic_decode::{
+    compact_u128, compact_u16, compact_u64, decode_compact_u8_vec,
+    extract_account_id_field, extract_balance_u128, pad_or_check, walk_account_vec,
+    walk_bool_vec, walk_compact_u128_vec, walk_compact_u16_vec, walk_compact_u64_vec,
+};
 use crate::error::BttError;
 use crate::rpc;
 
@@ -551,24 +556,6 @@ fn parse_metagraph<C: Clone>(
     Ok(SubnetMetagraphResult { header, uids })
 }
 
-/// Decode a `Vec<Compact<u8>>` value as a lossy UTF-8 String. Used
-/// for `name` and `symbol` fields. Returns `None` if the field is
-/// missing or any element is not a decodable u8.
-fn decode_compact_u8_vec<C: Clone>(composite: &Value<C>, field: &str) -> Option<String> {
-    let v = composite.at(field)?;
-    let mut bytes = Vec::new();
-    let mut idx = 0usize;
-    while let Some(entry) = v.at(idx) {
-        idx += 1;
-        let b = compact_value_to_u128(entry)?;
-        if b > 255 {
-            return None;
-        }
-        bytes.push(b as u8);
-    }
-    Some(String::from_utf8_lossy(&bytes).into_owned())
-}
-
 // ── hyperparameters ────────────────────────────────────────────────
 
 /// Full subnet hyperparameter dump returned by `btt subnet hyperparameters --netuid <N>`.
@@ -721,223 +708,13 @@ fn parse_hyperparams<C: Clone>(
     })
 }
 
-// ── per-UID vec walkers ───────────────────────────────────────────
-
-/// Normalize a per-UID Vec against the expected length. Three cases:
-/// - `got == expected`: pass through unchanged.
-/// - `got == 0` and `expected > 0`: backfill with `T::default()` to
-///   `expected` elements. This accommodates the runtime returning
-///   empty Vecs for fields whose storage has not yet been populated
-///   (e.g. `rank`, `trust`, `pruning_score` on a subnet whose epoch
-///   has not yet computed these values). The parallel-array indexing
-///   stays valid and the output contains zero placeholders which
-///   correctly represent the "not yet computed" state.
-/// - any other mismatch: hard error. A non-zero length that still
-///   doesn't match `num_uids` is runtime-version drift worth catching.
-fn pad_or_check<T: Default + Clone>(
-    got: Vec<T>,
-    expected: usize,
-    field: &str,
-) -> Result<Vec<T>, BttError> {
-    if got.len() == expected {
-        Ok(got)
-    } else if got.is_empty() {
-        Ok(vec![T::default(); expected])
-    } else {
-        Err(BttError::parse(format!(
-            "metagraph: per-UID array `{field}` has {got_len} elements, expected {expected} (num_uids) or 0",
-            got_len = got.len()
-        )))
-    }
-}
-
-fn walk_account_vec<C: Clone>(mg: &Value<C>, field: &str) -> Result<Vec<[u8; 32]>, BttError> {
-    let v = mg.at(field).ok_or_else(|| {
-        BttError::parse(format!("metagraph: missing vec field `{field}`"))
-    })?;
-    let mut out = Vec::new();
-    let mut idx = 0usize;
-    while let Some(entry) = v.at(idx) {
-        idx += 1;
-        let bytes = {
-            // Try the wrapped-tuple-struct shape first (AccountId32
-            // as a single-field composite wrapping [u8; 32]) and
-            // fall back to the raw byte array.
-            if let Some(inner) = entry.at(0) {
-                value_to_32_bytes(inner)
-            } else {
-                value_to_32_bytes(entry)
-            }
-        }
-        .ok_or_else(|| {
-            BttError::parse(format!(
-                "metagraph: `{field}`[{}] is not a decodable 32-byte account id",
-                idx - 1
-            ))
-        })?;
-        out.push(bytes);
-    }
-    Ok(out)
-}
-
-fn walk_compact_u16_vec<C: Clone>(mg: &Value<C>, field: &str) -> Result<Vec<u16>, BttError> {
-    walk_compact_numeric_vec(mg, field, |n| u16::try_from(n).ok())
-}
-
-fn walk_compact_u64_vec<C: Clone>(mg: &Value<C>, field: &str) -> Result<Vec<u64>, BttError> {
-    walk_compact_numeric_vec(mg, field, |n| u64::try_from(n).ok())
-}
-
-fn walk_compact_u128_vec<C: Clone>(mg: &Value<C>, field: &str) -> Result<Vec<u128>, BttError> {
-    walk_compact_numeric_vec(mg, field, Some)
-}
-
-fn walk_compact_numeric_vec<C: Clone, T>(
-    mg: &Value<C>,
-    field: &str,
-    coerce: impl Fn(u128) -> Option<T>,
-) -> Result<Vec<T>, BttError> {
-    let v = mg.at(field).ok_or_else(|| {
-        BttError::parse(format!("metagraph: missing vec field `{field}`"))
-    })?;
-    let mut out = Vec::new();
-    let mut idx = 0usize;
-    while let Some(entry) = v.at(idx) {
-        idx += 1;
-        let n = compact_value_to_u128(entry).ok_or_else(|| {
-            BttError::parse(format!(
-                "metagraph: `{field}`[{}] is not a decodable integer",
-                idx - 1
-            ))
-        })?;
-        let typed = coerce(n).ok_or_else(|| {
-            BttError::parse(format!(
-                "metagraph: `{field}`[{}] value {n} out of range for target type",
-                idx - 1
-            ))
-        })?;
-        out.push(typed);
-    }
-    Ok(out)
-}
-
-fn walk_bool_vec<C: Clone>(mg: &Value<C>, field: &str) -> Result<Vec<bool>, BttError> {
-    let v = mg.at(field).ok_or_else(|| {
-        BttError::parse(format!("metagraph: missing vec field `{field}`"))
-    })?;
-    let mut out = Vec::new();
-    let mut idx = 0usize;
-    while let Some(entry) = v.at(idx) {
-        idx += 1;
-        // scale-value represents booleans as Primitive::Bool under
-        // ValueDef::Primitive. Access via the primitive converter.
-        let b = entry.as_bool().ok_or_else(|| {
-            BttError::parse(format!(
-                "metagraph: `{field}`[{}] is not a bool",
-                idx - 1
-            ))
-        })?;
-        out.push(b);
-    }
-    Ok(out)
-}
-
-// ── decoder helpers ────────────────────────────────────────────────────
-//
-// These helpers intentionally duplicate the shape of the helpers in
-// `stake.rs`. Consolidating them into a shared module under
-// `src/commands/` is tracked by issue #93 as a pure refactor PR;
-// doing it inline here would creep scope across two commands.
-
-/// Extract a `Compact<u*>` field from a composite, coerced to u16.
-/// Fails if the field is missing, not a primitive, or exceeds u16 range.
-fn compact_u16<C: Clone>(composite: &Value<C>, field: &str) -> Option<u16> {
-    let v = composite.at(field)?;
-    compact_value_to_u128(v).and_then(|n| u16::try_from(n).ok())
-}
-
-/// Same, coerced to u64.
-fn compact_u64<C: Clone>(composite: &Value<C>, field: &str) -> Option<u64> {
-    let v = composite.at(field)?;
-    compact_value_to_u128(v).and_then(|n| u64::try_from(n).ok())
-}
-
-/// Same, left as u128 (for balances).
-fn compact_u128<C: Clone>(composite: &Value<C>, field: &str) -> Option<u128> {
-    let v = composite.at(field)?;
-    compact_value_to_u128(v)
-}
-
-/// The SCALE-value representation of a `Compact<T>` is a single-field
-/// composite wrapping the inner primitive. Walk the wrapper if
-/// present, otherwise read the primitive directly — both shapes show
-/// up in the wild depending on how the codec flattens on that runtime
-/// version.
-fn compact_value_to_u128<C: Clone>(v: &Value<C>) -> Option<u128> {
-    if let Some(n) = v.as_u128() {
-        return Some(n);
-    }
-    if let subxt::ext::scale_value::ValueDef::Composite(c) = &v.value {
-        let values: Vec<&Value<C>> = c.values().collect();
-        if values.len() == 1 {
-            return compact_value_to_u128(values[0]);
-        }
-    }
-    None
-}
-
-/// Pull a 32-byte AccountId out of a named field. Mirrors the helper
-/// in `stake.rs`; see the note above about consolidation.
-fn extract_account_id_field<C: Clone>(entry: &Value<C>, field: &str) -> Option<[u8; 32]> {
-    let field_val = entry.at(field)?;
-    if let Some(inner) = field_val.at(0) {
-        if let Some(bytes) = value_to_32_bytes(inner) {
-            return Some(bytes);
-        }
-    }
-    value_to_32_bytes(field_val)
-}
-
-fn value_to_32_bytes<C: Clone>(value: &Value<C>) -> Option<[u8; 32]> {
-    let mut bytes = Vec::with_capacity(32);
-    let mut idx = 0usize;
-    while let Some(v) = value.at(idx) {
-        let b = v.as_u128()?;
-        if b > 255 {
-            return None;
-        }
-        bytes.push(b as u8);
-        idx += 1;
-        if bytes.len() > 32 {
-            return None;
-        }
-    }
-    if bytes.len() == 32 {
-        let mut arr = [0u8; 32];
-        arr.copy_from_slice(&bytes);
-        Some(arr)
-    } else {
-        None
-    }
-}
-
-/// Walk a decoded `scale_value::Value` looking for a u128-compatible
-/// integer. Accepts either the bare-integer shape or a single-field
-/// composite wrapping one (the SCALE-Value codec occasionally produces
-/// the latter for tuple-struct newtypes).
-fn extract_balance_u128<C>(v: &subxt::dynamic::Value<C>) -> Option<u128> {
-    if let Some(n) = v.as_u128() {
-        return Some(n);
-    }
-    // Composite case: exactly one field, recurse.
-    if let subxt::ext::scale_value::ValueDef::Composite(c) = &v.value {
-        let values: Vec<&subxt::dynamic::Value<C>> = c.values().collect();
-        if values.len() == 1 {
-            return extract_balance_u128(values[0]);
-        }
-    }
-    None
-}
+// Decoder helpers (compact_u16/u64/u128, compact_value_to_u128,
+// extract_account_id_field, value_to_32_bytes, extract_balance_u128,
+// decode_compact_u8_vec, pad_or_check, walk_account_vec,
+// walk_compact_u16_vec, walk_compact_u64_vec, walk_compact_u128_vec,
+// walk_compact_numeric_vec, walk_bool_vec) moved to
+// `crate::commands::dynamic_decode` in issue #93. Imported at the top
+// of this file.
 
 /// Format a rao integer (u128) as a decimal TAO string with 9 fractional
 /// digits. The string representation is stable across hosts — no
